+++ conf/proxy.conf
@@ -233,7 +233,7 @@ clusterName=
 
 ### --- Proxy Extensions
 
-# List of proxy extensions to load, which is a list of protocol names
+# List of proxy extensions to load, which is a list of extension names
 #proxyExtensions=
 
 # The directory to locate extensions
@@ -243,4 +243,3 @@ clusterName=
 
 # Deprecated. Use configurationStoreServers
 globalZookeeperServers=
-
